backend/requirements.txt
asgiref==3.9.1
Cython==3.1.2
dj-database-url==3.0.1
Django==5.1.11
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
whitenoise==6.9.0
